requirements.txt
colorama==0.4.6
pillow==9.3.0
pyfiglet==0.8.post1